# openapi.yaml
openapi: 3.0.3
info:
  title: LifeBalance Sprint 3 Resource Capital and Task Core API
  version: 1.0.0
  description: >
    OpenAPI contract for LB-828 covering Sprint 3 Resource Capital Foundation
    and Task Core endpoints. All documented responses use the LifeBalance
    ApiResponse envelope: success, data, error, and timestamp.
  contact:
    name: LifeBalance Team
    email: team@lifebalance.com
  license:
    name: LifeBalance
    url: https://lifebalance.com/license
servers:
  - url: /
    description: Current service or gateway host
security:
  - bearerAuth: []
...
paths:
...
  /api/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get task by id
      description: Get a single task by id.
      operationId: getTaskById
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Tasks
      summary: Update task
      description: Update task details, progress, and status.
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            example:
              name: Prepare sprint planning
              description: Finalize scope and update task estimates.
              priority: HIGH
              deadline: '2026-08-20'
              progress: 50
              estimatedMinutes: 150
              estimatedCost: 0
              categoryId: 55555555-5555-5555-5555-555555555555
              status: IN_PROGRESS
      responses:
        '200':
          description: Task updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task by id.
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token using the Bearer authentication scheme.
  parameters:
...
    TaskIdPath:
      name: id
      in: path
      required: true
      description: Task id.
      schema:
        type: string
        format: uuid
      example: 77777777-7777-7777-7777-777777777777
...
  responses:
    BadRequest:
      description: Bad Request or validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: COMMON_VALIDATION_FAILED
              message: Request validation failed
              details:
                name: must not be blank
            timestamp: '2026-08-14T00:00:00Z'
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: AUTH_UNAUTHORIZED
              message: Authentication is required
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    Forbidden:
      description: Authenticated user does not have permission to access the resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: AUTH_FORBIDDEN
              message: Access is denied
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: COMMON_NOT_FOUND
              message: Resource was not found
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    Conflict:
      description: Business conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: RESOURCE_CONFLICT
              message: Business rule conflict
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: COMMON_INTERNAL_ERROR
              message: Unexpected server error
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
  schemas:
    ApiResponseBase:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
          example: '2026-08-14T00:00:00Z'
    ApiResponseError:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required:
            - success
            - error
            - timestamp
          properties:
            success:
              type: boolean
              example: false
            data:
              type: object
              nullable: true
              description: Always null for failed responses.
            error:
              $ref: '#/components/schemas/ApiError'
    ApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required:
            - success
            - timestamp
          properties:
            data:
              type: object
              nullable: true
              description: Null for void success responses.
            error:
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/ApiError'
...
    ApiResponseTask:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/TaskResponse'
            error:
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/ApiError'
...
    ApiError:
      type: object
      required:
        - code
        - message
        - details
      properties:
        code:
          type: string
          example: COMMON_VALIDATION_FAILED
        message:
          type: string
          example: Request validation failed
        details:
          type: object
          additionalProperties:
            type: string
          example:
            name: must not be blank
...
    UpdateTaskRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Prepare sprint planning
        description:
          type: string
          nullable: true
          example: Finalize scope and update task estimates.
        priority:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PriorityLevel'
        deadline:
          type: string
          format: date
          nullable: true
          example: '2026-08-20'
        progress:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          maximum: 100
          example: 50
        estimatedMinutes:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          example: 150
        estimatedCost:
          type: number
          format: decimal
          nullable: true
          minimum: 0
          example: 0
        categoryId:
          type: string
          format: uuid
          nullable: true
          example: 55555555-5555-5555-5555-555555555555
        status:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/TaskStatus'
    TaskResponse:
      type: object
      required:
        - id
        - userId
        - name
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: 77777777-7777-7777-7777-777777777777
        userId:
          type: string
          format: uuid
          example: 11111111-1111-1111-1111-111111111111
        name:
          type: string
          example: Prepare sprint planning
        description:
          type: string
          nullable: true
          example: Review capacity and scope for Sprint 3.
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PriorityLevel'
        deadline:
          type: string
          format: date
          nullable: true
          example: '2026-08-20'
        progress:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          maximum: 100
          example: 50
        estimatedMinutes:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          example: 120
        estimatedCost:
          type: number
          format: decimal
          nullable: true
          minimum: 0
          example: 0
        categoryId:
          type: string
          format: uuid
          nullable: true
          example: 55555555-5555-5555-5555-555555555555
        categoryName:
          type: string
          nullable: true
          example: Work
        createdAt:
          type: string
          format: date-time
          example: '2026-08-14T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-08-14T00:00:00Z'
...
    TaskStatus:
      type: string
      enum:
        - DRAFT
        - PLANNED
        - SCHEDULED
        - IN_PROGRESS
        - ON_HOLD
        - COMPLETED
        - CANCELLED
        - ARCHIVED
      example: IN_PROGRESS
    PriorityLevel:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL
      example: HIGH